% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{compute_covariance}
\alias{compute_covariance}
\title{Calculate Coefficient of Variation}
\usage{
compute_covariance(a, b)
}
\arguments{
\item{a}{The first numerical vector.}

\item{b}{The second numerical vector.}
}
\value{
A numeric value representing the coefficient of variation.
}
\description{
Calculates the coefficient of variation (CV) between two numerical vectors.
}
\examples{
data1 <- c(1, 2, 5, NA, 10)
data2 <- c(2, 4, 5, 8, NA)
compute_covariance(data1, data2)
}
