% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{covariance_statistics}
\alias{covariance_statistics}
\title{Calculate Validation Statistics with Bootstrapping}
\usage{
covariance_statistics(
  data,
  species_variable,
  growth_function,
  death_function,
  Nbootstraps = 500
)
}
\arguments{
\item{data}{A data frame containing the variables needed for the calculations.}

\item{species_variable}{Species variable in the data frame (provided as a string).}

\item{growth_function}{Expression of the growth function (provided as a string).}

\item{death_function}{Expression of the death function (provided as a string).}

\item{Nbootstraps}{The number of bootstrap iterations to perform (default is 500).}
}
\value{
A data frame containing z-scores of the bootstrapped coefficients of variation (CoV)
  for growth and death.
}
\description{
Calculates bootstrapped validation statistics (coefficients of variation)
  using user-provided growth and death functions.
}
\examples{
set.seed(1010)
covariance_statistics(
  example_data,
  species_variable = "prey",
  growth_function = "prey",
  death_function = "prey*predator + prey^2"
)
}
