% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{An example dataset of ecological time series}
\format{
## `example_data`
A data frame with 198 row and 3 columns:
\describe{
  \item{time}{Measurement time}
  \item{prey}{Prey abundance}
  \item{predator}{Predator abundance}
}
}
\source{
<https://www.nature.com/articles/s41586-019-1857-0>
}
\usage{
example_data
}
\description{
An example dataset of ecological time series
}
\keyword{datasets}
